Add button editing classes
