+++ app_pojavlauncher/src/main/java/net/kdt/pojavlaunch/customcontrols/handleview/ActionPopupWindow.java
@@ -21,8 +21,8 @@ package net.kdt.pojavlaunch.customcontrols.handleview;
 
 import android.app.Dialog;
 import android.content.*;
-import android.graphics.Color;
 import android.graphics.drawable.ColorDrawable;
+import android.util.Log;
 import android.view.*;
 import android.view.ViewGroup.*;
 import android.widget.*;
@@ -36,48 +36,15 @@ import androidx.appcompat.app.*;
 import com.rarepebble.colorpicker.ColorPickerView;
 
 public class ActionPopupWindow extends PinnedPopupWindow implements OnClickListener {
+
 	private TextView mEditTextView;
 	private TextView mDeleteTextView;
 
-	private AlertDialog.Builder alertBuilder;
-	private Dialog dialog;
+	private Object editedButton;
 
-	private EditText editName;
-	private Spinner[] spinnersKeycode;
-
-	private CheckBox checkToggle;
-	private CheckBox checkPassThrough;
-	private CheckBox checkDynamicPosition;
-	private CheckBox checkHoldAlt;
-	private CheckBox checkHoldCtrl;
-	private CheckBox checkHoldShift;
-
-	private EditText editWidth;
-	private EditText editHeight;
-	private EditText editDynamicX;
-	private EditText editDynamicY;
-
-	private SeekBar seekBarOpacity;
-	private SeekBar seekBarCornerRadius;
-	private SeekBar seekBarStrokeWidth;
-
-	private ImageButton buttonBackgroundColor;
-	private ImageButton buttonStrokeColor;
-
-	private TextView textOpacity;
-	private TextView textCornerRadius;
-	private TextView textStrokeWidth;
-
-	private ControlData properties;
-
-	private ArrayAdapter<String> adapter;
-	private String[] specialArr;
-
-
-
-
-	public ActionPopupWindow(HandleView handleView) {
+	public ActionPopupWindow(HandleView handleView, Object object){
 		super(handleView);
+		this.editedButton = object;
 	}
 
 	@Override
@@ -119,26 +86,46 @@ public class ActionPopupWindow extends PinnedPopupWindow implements OnClickListe
 
 	@Override
 	public void onClick(final View view) {
-		alertBuilder = new AlertDialog.Builder(view.getContext());
-		alertBuilder.setCancelable(false);
-
-
-
-
 		if (view == mEditTextView) {
-			properties = mHandleView.mView.getProperties();
-			initializeEditDialog(view.getContext());
+
+			if(editedButton instanceof ControlSubButton){
+				new EditControlSubButtonPopup((ControlSubButton) editedButton);
+				return;
+			}
+
+			if(editedButton instanceof ControlDrawer){
+				new EditControlDrawerPopup((ControlDrawer) editedButton);
+				return;
+			}
+
+			if(editedButton instanceof ControlButton){
+				new EditControlButtonPopup((ControlButton) editedButton);
+				return;
+			}
 
 
-
-
-
-			dialog.show();
 		} else if (view == mDeleteTextView) {
+			AlertDialog.Builder alertBuilder = new AlertDialog.Builder(view.getContext());
+			alertBuilder.setCancelable(false);
 			alertBuilder.setMessage(view.getContext().getString(R.string.customctrl_remove, mHandleView.mView.getText()) + "?");
 
 			alertBuilder.setPositiveButton(R.string.global_remove, (p1, p2) -> {
 				ControlLayout layout = ((ControlLayout) mHandleView.mView.getParent());
+
+				if(editedButton instanceof ControlSubButton){
+					layout.removeControlSubButton((ControlSubButton) editedButton);
+					return;
+				}
+
+				if(editedButton instanceof ControlDrawer){
+					layout.removeControlDrawer((ControlDrawer) editedButton);
+					return;
+				}
+
+				if(editedButton instanceof ControlButton){
+					layout.removeControlButton((ControlButton) editedButton);
+				}
+
 				layout.removeControlButton(mHandleView.mView);
 			});
 			alertBuilder.setNegativeButton(android.R.string.cancel, null);
@@ -163,272 +150,22 @@ public class ActionPopupWindow extends PinnedPopupWindow implements OnClickListe
 		return positionY;
 	}
 
-	private void initializeEditDialog(Context ctx){
-		//Create the editing dialog
-		LayoutInflater layoutInflater = (LayoutInflater) ctx.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-		View v = layoutInflater.inflate(R.layout.control_setting_v2,null);
 
-		alertBuilder.setTitle(ctx.getResources().getString(R.string.customctrl_edit, mHandleView.mView.getText()));
-		alertBuilder.setView(v);
-
-		//Linking a lot of stuff
-		editName = v.findViewById(R.id.controlsetting_edit_name);
-
-		spinnersKeycode = new Spinner[]{
-				v.findViewById(R.id.controlsetting_spinner_lwjglkeycode),
-				v.findViewById(R.id.controlsetting_spinner_lwjglkeycode2),
-				v.findViewById(R.id.controlsetting_spinner_lwjglkeycode3),
-				v.findViewById(R.id.controlsetting_spinner_lwjglkeycode4)
-		};
-		checkToggle = v.findViewById(R.id.controlsetting_checkbox_toggle);
-		checkPassThrough = v.findViewById(R.id.controlsetting_checkbox_passthru);
-
-		editWidth = v.findViewById(R.id.controlsetting_edit_width);
-		editHeight = v.findViewById(R.id.controlsetting_edit_height);
-
-		editDynamicX = v.findViewById(R.id.controlsetting_edit_dynamicpos_x);
-		editDynamicY = v.findViewById(R.id.controlsetting_edit_dynamicpos_y);
-
-		seekBarOpacity = v.findViewById(R.id.controlsetting_seek_opacity);
-		seekBarCornerRadius = v.findViewById(R.id.controlsetting_seek_corner_radius);
-		seekBarStrokeWidth = v.findViewById(R.id.controlsetting_seek_stroke_width);
-
-		//Add listeners, too bad I don't need all the methods
-		seekBarOpacity.setOnSeekBarChangeListener(new SeekBar.OnSeekBarChangeListener() {
-			@Override
-			public void onProgressChanged(SeekBar seekBar, int i, boolean b) {
-				setPercentageText(textOpacity, i);
-			}
-
-			@Override
-			public void onStartTrackingTouch(SeekBar seekBar) {
-				//AUTO GENERATED STUN
-			}
-			@Override
-			public void onStopTrackingTouch(SeekBar seekBar) {
-				//AUTO GENERATED STUN
-			}
-		});
-
-		seekBarCornerRadius.setOnSeekBarChangeListener(new SeekBar.OnSeekBarChangeListener() {
-			@Override
-			public void onProgressChanged(SeekBar seekBar, int i, boolean b) {
-				setPercentageText(textCornerRadius, i);
-			}
-
-			@Override
-			public void onStartTrackingTouch(SeekBar seekBar) {
-				//AUTO GENERATED STUN
-			}
-			@Override
-			public void onStopTrackingTouch(SeekBar seekBar) {
-				//AUTO GENERATED STUN
-			}
-		});
-
-		seekBarStrokeWidth.setOnSeekBarChangeListener(new SeekBar.OnSeekBarChangeListener() {
-			@Override
-			public void onProgressChanged(SeekBar seekBar, int i, boolean b) {
-				setPercentageText(textStrokeWidth, i);
-			}
-
-			@Override
-			public void onStartTrackingTouch(SeekBar seekBar) {
-				//AUTO GENERATED STUN
-			}
-			@Override
-			public void onStopTrackingTouch(SeekBar seekBar) {
-				//AUTO GENERATED STUN
-			}
-		});
-
-		buttonBackgroundColor = v.findViewById(R.id.controlsetting_background_color);
-		buttonStrokeColor = v.findViewById(R.id.controlsetting_stroke_color);
-
-		textOpacity = v.findViewById(R.id.controlsetting_text_opacity);
-		textCornerRadius = v.findViewById(R.id.controlsetting_text_corner_radius);
-		textStrokeWidth = v.findViewById(R.id.controlsetting_text_stroke_width);
-
-		checkDynamicPosition = v.findViewById(R.id.controlsetting_checkbox_dynamicpos);
-		checkDynamicPosition.setOnCheckedChangeListener((btn, checked) -> {
-			editDynamicX.setEnabled(checked);
-			editDynamicY.setEnabled(checked);
-		});
-
-
-		checkHoldAlt = v.findViewById(R.id.controlsetting_checkbox_keycombine_alt);
-		checkHoldCtrl = v.findViewById(R.id.controlsetting_checkbox_keycombine_control);
-		checkHoldShift = v.findViewById(R.id.controlsetting_checkbox_keycombine_shift);
-
-		//Initialize adapter for keycodes
-		adapter= new ArrayAdapter<String>(ctx, android.R.layout.simple_spinner_item);
-		String[] oldSpecialArr = ControlData.buildSpecialButtonArray();
-		specialArr = new String[oldSpecialArr.length];
-		for (int i = 0; i < specialArr.length; i++) {
-			specialArr[i] = "SPECIAL_" + oldSpecialArr[specialArr.length - i - 1];
-		}
-		adapter.addAll(specialArr);
-		adapter.addAll(AndroidLWJGLKeycode.generateKeyName());
-		adapter.setDropDownViewResource(android.R.layout.simple_list_item_single_choice);
-
-		for (Spinner spinner : spinnersKeycode) {
-			spinner.setAdapter(adapter);
-		}
-
-		//Set color imageButton behavior
-		buttonBackgroundColor.setOnClickListener(view -> showBackgroundColorPicker(view.getContext()));
-		buttonStrokeColor.setOnClickListener(view -> showStrokeColorPicker(view.getContext()));
-
-		//Set dialog buttons behavior
-		alertBuilder.setPositiveButton(android.R.string.ok, (dialogInterface1, i) -> {
-			if(!hasErrors(dialog.getContext())){
-				saveProperties();
-			}
-		});
-		alertBuilder.setNegativeButton(android.R.string.cancel, null);
-
-		//Create the finalized dialog
-		dialog = alertBuilder.create();
-
-		dialog.setOnShowListener(dialogInterface -> {
-			setEditDialogValues();
-		});
-
-	}
-
-	private void setEditDialogValues(){
-
-		editName.setText(properties.name);
-
-		checkToggle.setChecked(properties.isToggle);
-		checkPassThrough.setChecked(properties.passThruEnabled);
-
-		editWidth.setText(Float.toString(properties.width));
-		editHeight.setText(Float.toString(properties.height));
-
-		editDynamicX.setEnabled(properties.isDynamicBtn);
-		editDynamicY.setEnabled(properties.isDynamicBtn);
-		editDynamicX.setHint(Float.toString(properties.x));
-		editDynamicX.setText(properties.dynamicX);
-		editDynamicY.setHint(Float.toString(properties.y));
-		editDynamicY.setText(properties.dynamicY);
-
-		seekBarOpacity.setProgress(properties.hidden ? 0 : (int)properties.opacity*100);
-		seekBarStrokeWidth.setProgress(properties.strokeWidth);
-		seekBarCornerRadius.setProgress((int)properties.cornerRadius);
-
-		buttonBackgroundColor.setBackgroundColor(properties.bgColor);
-		buttonStrokeColor.setBackgroundColor(properties.strokeColor);
-
-		setPercentageText(textCornerRadius,seekBarCornerRadius.getProgress());
-		setPercentageText(textOpacity,seekBarOpacity.getProgress());
-		setPercentageText(textStrokeWidth,seekBarStrokeWidth.getProgress());
-
-		checkHoldAlt.setChecked(properties.holdAlt);
-		checkHoldCtrl.setChecked(properties.holdCtrl);
-		checkHoldShift.setChecked(properties.holdShift);
-		checkDynamicPosition.setChecked(properties.isDynamicBtn);
-
-		for(int i=0; i< properties.keycodes.length; i++){
-			if (properties.keycodes[i] < 0) {
-				spinnersKeycode[i].setSelection(properties.keycodes[i] + specialArr.length);
-			} else {
-				spinnersKeycode[i].setSelection(AndroidLWJGLKeycode.getIndexByLWJGLKey(properties.keycodes[i]) + specialArr.length);
-			}
-		}
-	}
-
-	private void saveProperties(){
-		//TODO save new properties
-		//This method assumes there are no error.
-		properties.name = editName.getText().toString();
-
-		//Keycodes
-		for(int i=0; i<spinnersKeycode.length; ++i){
-			if (spinnersKeycode[i].getSelectedItemPosition() < specialArr.length) {
-				properties.keycodes[i] = spinnersKeycode[i].getSelectedItemPosition() - specialArr.length;
-			} else {
-				properties.keycodes[i] = AndroidLWJGLKeycode.getKeyByIndex(spinnersKeycode[i].getSelectedItemPosition() - specialArr.length);
-			}
-		}
-
-		properties.opacity = seekBarOpacity.getProgress()/100f;
-		properties.strokeWidth = seekBarStrokeWidth.getProgress();
-		properties.cornerRadius = seekBarCornerRadius.getProgress();
-
-		properties.bgColor = ((ColorDrawable) buttonBackgroundColor.getBackground()).getColor();
-		properties.strokeColor = ((ColorDrawable) buttonStrokeColor.getBackground()).getColor();
-
-		properties.hidden = false;
-		properties.isToggle = checkToggle.isChecked();
-		properties.passThruEnabled = checkPassThrough.isChecked();
-
-		properties.width = Float.parseFloat(editWidth.getText().toString());
-		properties.height = Float.parseFloat(editHeight.getText().toString());
-
-		properties.isDynamicBtn = checkDynamicPosition.isChecked();
-		properties.dynamicX = editDynamicX.getText().toString().isEmpty() ? properties.dynamicX = Float.toString(properties.x) : editDynamicX.getText().toString();
-		properties.dynamicY = editDynamicY.getText().toString().isEmpty() ? properties.dynamicY = Float.toString(properties.y) : editDynamicY.getText().toString();
-
-		properties.holdAlt = checkHoldAlt.isChecked();
-		properties.holdCtrl = checkHoldCtrl.isChecked();
-		properties.holdShift = checkHoldShift.isChecked();
-
-		mHandleView.mView.updateProperties();
-	}
-
-	private boolean hasErrors(Context ctx){
-		if (editName.getText().toString().isEmpty()) {
-			editName.setError(ctx.getResources().getString(R.string.global_error_field_empty));
-			return true;
-		}
-
-		if (properties.isDynamicBtn) {
-
-			int errorAt = 0;
-			try {
-				properties.insertDynamicPos(editDynamicX.getText().toString());
-				errorAt = 1;
-				properties.insertDynamicPos(editDynamicY.getText().toString());
-			} catch (Throwable th) {
-				(errorAt == 0 ? editDynamicX : editDynamicY).setError(th.getMessage());
-
-				return true;
-			}
-		}
-
-		return false;
-	}
-
-
-	private void showBackgroundColorPicker(Context ctx){
+	public static void showColorPicker(Context ctx,String title, boolean showAlpha, int startColor, View v){
 		ColorPickerView picker = new ColorPickerView(ctx);
-		picker.setColor(((ColorDrawable)buttonBackgroundColor.getBackground()).getColor());
+		picker.setColor(startColor);
+		picker.showAlpha(showAlpha);
 
 		AlertDialog.Builder dialog = new AlertDialog.Builder(ctx);
-		dialog.setTitle("Edit background color");
+		dialog.setTitle(title);
 		dialog.setView(picker);
 		dialog.setNegativeButton(android.R.string.cancel, null);
-		dialog.setPositiveButton(android.R.string.ok, (dialogInterface, i) -> buttonBackgroundColor.setBackgroundColor(picker.getColor()));
+		dialog.setPositiveButton(android.R.string.ok, (dialogInterface, i) -> v.setBackgroundColor(picker.getColor()));
 
 		dialog.show();
 	}
 
-	private void showStrokeColorPicker(Context ctx){
-		ColorPickerView picker = new ColorPickerView(ctx);
-		picker.setColor(((ColorDrawable)buttonStrokeColor.getBackground()).getColor());
-		picker.showAlpha(false);
-
-		AlertDialog.Builder dialog = new AlertDialog.Builder(ctx);
-		dialog.setTitle("Edit stroke color");
-		dialog.setView(picker);
-		dialog.setNegativeButton(android.R.string.cancel, null);
-		dialog.setPositiveButton(android.R.string.ok, (dialogInterface, i) -> buttonStrokeColor.setBackgroundColor(picker.getColor()));
-
-		dialog.show();
-	}
-
-	private void setPercentageText(TextView textView, int progress){
+	public static void setPercentageText(TextView textView, int progress){
 		textView.setText(progress + " %");
 	}
